requirements.txt
mysql-connector-python==8.0.29
protobuf==4.21.1
python-dotenv==0.20.0
